+++ packages/claudexor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # claudexor
 
+## 3.9.8
+
+### Patch Changes
+
+- @claudexor/cli@3.9.8
+
 ## 3.9.7
 
 ### Patch Changes