+++ packages/control-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @claudexor/control-api
 
+## 3.9.8
+
+### Patch Changes
+
+- @claudexor/delivery@3.9.8
+- @claudexor/event-log@3.9.8
+- @claudexor/schema@3.9.8
+- @claudexor/secrets@3.9.8
+- @claudexor/util@3.9.8
+
 ## 3.9.7
 
 ### Patch Changes