+++ packages/delivery/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @claudexor/delivery
 
+## 3.9.8
+
+### Patch Changes
+
+- @claudexor/core@3.9.8
+- @claudexor/policy@3.9.8
+- @claudexor/review@3.9.8
+- @claudexor/schema@3.9.8
+- @claudexor/util@3.9.8
+- @claudexor/workspace@3.9.8
+
 ## 3.9.7
 
 ### Patch Changes