+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @claudexor/gateway
 
+## 3.9.8
+
+### Patch Changes
+
+- @claudexor/core@3.9.8
+- @claudexor/schema@3.9.8
+
 ## 3.9.7
 
 ### Patch Changes