+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/schema
 
+## 3.9.8
+
+### Patch Changes
+
+- @claudexor/util@3.9.8
+
 ## 3.9.7
 
 ### Patch Changes